# Conflicts:
#	integration-tests/src/test/java/cc/blynk/integration/http/HttpAPIPinsAsyncClientTest.java
#	integration-tests/src/test/java/cc/blynk/integration/tcp/DeviceTilesWidgetTest.java
#	integration-tests/src/test/java/cc/blynk/integration/tcp/SyncWorkflowTest.java
#	server/core/src/main/java/cc/blynk/server/core/dao/ReportingDiskDao.java
#	server/core/src/main/java/cc/blynk/server/core/model/DashBoard.java
#	server/core/src/main/java/cc/blynk/server/core/model/DataStream.java
#	server/core/src/main/java/cc/blynk/server/core/reporting/average/AggregationKey.java
#	server/core/src/main/java/cc/blynk/server/core/reporting/raw/BaseReportingKey.java
#	server/core/src/main/java/cc/blynk/server/db/dao/ReportingDBDao.java
#	server/core/src/test/java/cc/blynk/server/core/model/DataStreamValuesUpdateCorrectTest.java
#	server/core/src/test/java/cc/blynk/server/core/model/widgets/DataStreamGetJsonValueTest.java
#	server/core/src/test/java/cc/blynk/server/core/model/widgets/MultiPinWidgetsToJsonTest.java
#	server/core/src/test/java/cc/blynk/server/db/RawDataDBTest.java
#	server/core/src/test/java/cc/blynk/server/db/RealtimeStatsDBTest.java
#	server/launcher/src/test/java/cc/blynk/server/workers/ReportingWorkerTest.java
#	server/tcp-app-server/src/main/java/cc/blynk/server/application/handlers/main/logic/MobileHardwareResendFromBTLogic.java

+++ server/core/src/main/java/cc/blynk/server/core/model/device/Device.java
@@ -248,7 +248,7 @@ public class Device implements Target {
         updateFromMobile(newDevice);
     }
 
-    public void updateWebDashboard(byte pin, PinType type, String value, long now) {
+    public void updateWebDashboard(short pin, PinType type, String value, long now) {
         webDashboard.update(id, pin, type, value);
         this.dataReceivedAt = now;
     }
